+++ app/javascript/stylesheets/booking.module.css
@@ -55,6 +55,7 @@
     color: #fff;
     line-height: 26px;
     cursor: pointer;
+    font-size: 16px;
 }
 
 .bookingButton:hover {
