+++ app/javascript/stylesheets/favorite_places.css
@@ -1,32 +1,35 @@
 .favorites { 
-    color: #173166;
-    font-family: 'Trebuchet MS', sans-serif;
+  color: #173166;
+  font-family: 'Trebuchet MS', sans-serif;
 
 }
 
 h1 {
-    margin: 10px;
-    text-align: center;
+  margin: 10px;
+  text-align: center;
 }
 
 .favorite-list{
-    color: white;
-    font-size: 20px;
-    margin: 0 auto;
-    max-width: 500px;
-    background-color: #173166;
-    padding: 25px;
-    border-radius: 4px;
-    box-shadow: rgba(1,60,136,.5) 0 -1px 3px 0 inset,rgba(0,44,97,.1) 0 3px 6px 0;
+  display: flex;
+  justify-content: space-between;
+  color: white;
+  font-size: 20px;
+  margin: 5px;
+  max-width: 500px;
+  background-color: #173166;
+  padding: 25px;
+  border-radius: 4px;
+  box-shadow: rgba(1,60,136,.5) 0 -1px 3px 0 inset,rgba(0,44,97,.1) 0 3px 6px 0;
 }
 
 ul {
-    display: flex;
-    justify-content: space-between;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
 }
 
 li {
-    margin: 5px;
+  margin: 5px;
 }
 
 .del_favorite {
@@ -58,29 +61,4 @@ li {
   background-color: #F3F4F6;
   text-decoration: none;
   transition-duration: 0.1s;
-}
-
-.button-4:disabled {
-  background-color: #FAFBFC;
-  border-color: rgba(27, 31, 35, 0.15);
-  color: #959DA5;
-  cursor: default;
-}
-
-.button-4:active {
-  background-color: #EDEFF2;
-  box-shadow: rgba(225, 228, 232, 0.2) 0 1px 0 inset;
-  transition: none 0s;
-}
-
-.button-4:focus {
-  outline: 1px transparent;
-}
-
-.button-4:before {
-  display: none;
-}
-
-.button-4:-webkit-details-marker {
-  display: none;
 }
