+++ app/serializers/place_serializer.rb
@@ -1,4 +1,3 @@
 class PlaceSerializer < ActiveModel::Serializer
-  attributes :id, :number, :room_id
-  has_many :favorites
+  attributes :id, :number, :room_id, :created_at, :updated_at
 end
