+++ package.json
@@ -3,7 +3,8 @@
     "@babel/preset-react": "^7.18.6",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
-    "@mui/material": "^5.10.12",
+    "@mui/icons-material": "^5.10.9",
+    "@mui/material": "^5.10.13",
     "@rails/webpacker": "5.4.3",
     "axios": "^0.27.2",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
